+++ cf/test/test_Data.py
@@ -73,30 +73,31 @@ class DataTest(unittest.TestCase):
         self.test_only = []
 #        self.test_only = ['NOTHING!!!!!']
 
-        self.test_only = ['test_Data_AUXILIARY_MASK',
-                          'test_Data_datum',
-#                          'test_Data_ERROR',
-                          'test_Data_array',
-                          'test_Data_varray',
-                          'test_Data_datetime_array',
-#                          'test_Data_cumsum',
-                          'test_Data_dumpd_loadd_dumps',
-#                          'test_Data_sin_cos_tan',
-#                          'test_Data_root_mean_square',
-#                          'test_Data_mean_mean_absolute_value',
-                          'test_Data_squeeze_insert_dimension',
-                          'test_Data_months_years',
-                          'test_Data_binary_mask',
-#                          'test_Data_CachedArray',
-                          'test_Data_digitize',
-#                          'test_Data_outerproduct',
-#                          'test_Data_flatten',
-#                          'test_Data_transpose',
-                          'test_Data__collapse_SHAPE',
-#                          'test_Data_range_mid_range',
-#                          'test_Data_median',
-#                          'test_Data_mean_of_upper_decile',
-        ]
+#        self.test_only = ['test_Data_AUXILIARY_MASK',
+#                          'test_Data_datum',
+##                          'test_Data_ERROR',
+#                          'test_Data_array',
+#                          'test_Data_varray',
+#                          'test_Data_datetime_array',
+##                          'test_Data_cumsum',
+#                          'test_Data_dumpd_loadd_dumps',
+##                          'test_Data_sin_cos_tan',
+##                          'test_Data_root_mean_square',
+##                          'test_Data_mean_mean_absolute_value',
+#                          'test_Data_squeeze_insert_dimension',
+#                          'test_Data_months_years',
+#                          'test_Data_binary_mask',
+##                          'test_Data_CachedArray',
+#                          'test_Data_digitize',
+##                          'test_Data_outerproduct',
+##                          'test_Data_flatten',
+##                          'test_Data_transpose',
+#                          'test_Data__collapse_SHAPE',
+##                          'test_Data_range_mid_range',
+##                          'test_Data_median',
+##                          'test_Data_mean_of_upper_decile',
+#                          'test_Data__init__mask',
+#        ]
         
 #        self.test_only = ['test_Data_mean_mean_absolute_value']
 #        self.test_only = ['test_Data_AUXILIARY_MASK']
@@ -108,7 +109,7 @@ class DataTest(unittest.TestCase):
 #        self.test_only = ['test_Data__collapse_WEIGHTED_MASKED']
 #        self.test_only = ['test_Data_ERROR']
 #        self.test_only = ['test_Data_sample_size']
-#        self.test_only = ['test_Data_section']
+#        self.test_only = ['test_Data__init__mask']
 #        self.test_only = ['test_Data_sd_var']
 #        self.test_only = ['test_Data_sum_of_weights_sum_of_weights2']
 #        self.test_only = ['test_Data_max_min_sum_sum_of_squares']
@@ -119,6 +120,35 @@ class DataTest(unittest.TestCase):
 #        self.test_only = ['test_Data_clip']
 
 
+    def test_Data__init__mask(self):
+        if self.test_only and inspect.stack()[0][3] not in self.test_only:
+            return
+
+        for m in (1, 20, True):
+            d = cf.Data([[1,2,3], [4,5,6]], mask=m)
+            self.assertTrue(not d.count())
+            self.assertTrue(d.shape == (2, 3))
+        
+        for m in (0, False):
+            d = cf.Data([[1,2,3], [4,5,6]], mask=m)
+            self.assertTrue(d.count() == d.size)
+            self.assertTrue(d.shape == (2, 3))
+            
+        
+        d = cf.Data([[1,2,3], [4,5,6]], mask=[[0], [1]])
+        self.assertTrue(d.count() == 3)
+        self.assertTrue(d.shape == (2, 3))
+            
+        d = cf.Data([[1,2,3], [4,5,6]], mask=[0, 1, 1])
+        self.assertTrue(d.count() == 2)
+        self.assertTrue(d.shape == (2, 3))
+            
+        
+        d = cf.Data([[1,2,3], [4,5,6]], mask=[[0, 1, 0], [1, 0, 1]])
+        self.assertTrue(d.count() == 3)
+        self.assertTrue(d.shape == (2, 3))
+            
+        
     def test_Data_digitize(self):
         if self.test_only and inspect.stack()[0][3] not in self.test_only:
             return